Support ability to register CSIDriver
With this change, driver-registrar can be used to register
the driver with Kubernetes by submitting a CSIDriver CRD
object.
